fix: dmanh sửa giao diện quản lý danh mục dịch vụ, dịch vụ; sửa hiển thị dịch vụ ngoài trang chính, sửa hiển thị các bài viết ngoài trang chính

+++ app/Http/Controllers/Admin/UserController.php
@@ -96,7 +96,7 @@ class UserController extends Controller
 
         if ($request->hasFile('avatar')) {
             $imagePath = public_path('storage/images/avatar/' . $user->avatar);
-            if (File::exists($imagePath)) {
+            if (File::exists($imagePath) && $user->avatar != 'avatar_default.png') {
                 File::delete($imagePath);
             }
